+++ rbx_types/src/basic_types.rs
@@ -215,6 +215,16 @@ impl Color3 {
     }
 }
 
+impl From<Color3uint8> for Color3 {
+    fn from(value: Color3uint8) -> Self {
+        Self {
+            r: value.r as f32 / 255.0,
+            g: value.g as f32 / 255.0,
+            b: value.b as f32 / 255.0,
+        }
+    }
+}
+
 /// Represents non-HDR colors, i.e. those whose individual color channels do not
 /// exceed 1. This type is used for serializing properties like
 /// [`BasePart.Color`][BasePart.Color], but is not exposed as a distinct type to
@@ -235,6 +245,16 @@ pub struct Color3uint8 {
 impl Color3uint8 {
     pub fn new(r: u8, g: u8, b: u8) -> Self {
         Self { r, g, b }
+    }
+}
+
+impl From<Color3> for Color3uint8 {
+    fn from(value: Color3) -> Self {
+        Self {
+            r: ((value.r.max(0.0).min(1.0)) * 255.0).round() as u8,
+            g: ((value.g.max(0.0).min(1.0)) * 255.0).round() as u8,
+            b: ((value.b.max(0.0).min(1.0)) * 255.0).round() as u8,
+        }
     }
 }