+++ rbx_xml/CHANGELOG.md
@@ -1,6 +1,7 @@
 # rbx_xml Changelog
 
 ## Unreleased
+* Added conversion from `Color3` to `Color3uint8` to address [rojo-rbx/rojo#443](https://github.com/rojo-rbx/rojo/issues/443).
 
 ## 0.12.1 (2021-07-02)
 * Upgraded to rbx\_dom\_weak 2.1.